Stage #1: Main page header
Description

You're working on a college newspaper called Hypernews. You are comfortable in the digital world, so you came up with the idea of broadcasting the news online and decided to create a website for this. Creating the Hypernews site will be a long road, so let's take our first step forward and make the first page. For now, you won't have actual news on the site, so until that happens you can respond to the users with the "Coming soon" message.

Objectives

Create the first handler for your server.

If you start the project on your computer with python manage.py runserver command, your server should respond with the text "Coming soon" at the address localhost:8000.