Stage #2: The news page
Description

Exciting: you got the first news material for your e-paper, and that news is just extraterrestrial! Now you are going to proudly show it on the site. For that purpose, you'll need to add a page to display news.

Objectives

You will store data in the JSON file; the path to this file is saved in the NEWS_JSON_PATH variable in the project settings.

Your JSON file should contain an array of objects with obligatory created, text, title, and link fields:

[
  {
    "created": "2020-02-22 16:40:00",
    "text": "A new star appeared in the sky.",
    "title": "The birth of the star",
    "link": 9234732
  },
  ...
]
The field created should have the format "%Y-%m-%d %H:%M:%S".

Each news page has a unique link. In our case, the link identifier should be a number. You can pick any number you want, just make sure that it is unique for every news material.