Stage #4: Creating new news
Description

You store your news data in the JSON file. Every time you want to add news, you have to edit this file, which is quite inconvenient. That's why you need to add the possibility to create news through your web interface.

Objectives

Create a new page to add news using a form. This form should send data to the /news/create/ address with the POST method to create a new record with two parameters, title and text. When the request is completed, the response should redirect you back to the main page.

If you start the application on your computer with the python manage.py runserver command, the page for creating news should be available at the localhost:8000/news/create/ address.

The page should also contain a link to the main page link with target="_blank" href="/news/".

The link in your storage should be a random number, but remember that it's unique for every news. Choose the current time for the created field, and do not forget to convert it to a string before saving to JSON.